django-1.11.26/requirements.txt
Django==1.11.26
mysqlclient==1.4.6
Pillow==6.2.1
django-uuslug==1.2.0
django-mdeditor==0.1.16
django-haystack==2.8.1
jieba==0.39
Whoosh==2.7.4
Markdown==3.1.1
simpleui==3.2
six==1.13.0
gunicorn==20.0.4


